% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Attributes}
\alias{Attributes}
\title{Available attributes/parameters to filter and select species}
\format{
A \code{list} with 11 elements: States, Biome, vegetationType,
lifeForm, habitat, nomenclaturalStatus, taxonomicStatus, Endemism, Origin,
Group, and Subgroup
}
\usage{
data(Attributes)
}
\description{
A list of dataset containing the available attributes to filter and select
species. The dataset is used internally by \code{get_attributes}
}
\keyword{datasets}
