% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bf_data}
\alias{bf_data}
\title{Brazilian Flora 2020 database - Version 393.387}
\format{
A \code{data.frame} with 45890 rows and 19 variables:
\describe{
\item{species}{Species names}
\item{acceptedName}{Accepted name of the species (NA when the name in
species is already an accepted name)}
\item{kingdom}{Kingdom to which species belongs (Plantae or Fungi)}
\item{Group}{Major group to which species belongs (Angiosperms,
Gymnosperms, Ferns and Lycophytes, Bryophytes, and Algae)}
\item{Subgroup}{Subgroup to which species belongs. Only available for
Bryophytes (Mosses, Hornworts, and Liverworts)}
\item{family}{Family to which species belongs}
\item{genus}{Genus to which species belongs}
\item{lifeForm}{Life form of the species (e.g: Tree, Herb, Shrub, etc.)}
\item{habitat}{Habitat type of the species (e.g., Terrestrial, Rupicolous,
Epiphytic, etc.)}
\item{Biome}{Biomes with confirmed occurrences of the species}
\item{States}{Federal states with confirmed occurrences of the species}
\item{vegetationType}{Vegetation types with confirmed occurrences of the
species}
\item{Origin}{Indicates whether the species is Native, Naturalized, or
Cultivated in Brazil}
\item{Endemism}{Indicates whether the species is Endemic or Non-endemic to
Brazil}
\item{taxonomicStatus}{Indicates the level of recognition and acceptance
of the species (Accepted or Synonym)}
\item{nomenclaturalStatus}{Indicates the legitimacy and validity of the
species name (Correct, Illegitimate, Uncertain_Application, etc.)}
\item{vernacularName}{Locally or culturally used name for the species}
\item{taxonRank}{Taxonomic rank (Species, Genus, Family, Order, etc). This
data contains only Species}
}
}
\usage{
bf_data
}
\description{
A dataset containing a subset of the Brazilian Flora database (version
393.399)

@usage data(bf_data)
}
\keyword{datasets}
