% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomes}
\alias{biomes}
\title{SpatVector of the biomes of Brazil}
\format{
A \code{SpatVector} with 6 geometries and 1 attribute:
\describe{
\item{name_biome}{The name of the biome (Amazon, Caatinga, Cerrado,
Atlantic_Forest, Pampa, and Pantanal)}
}
}
\usage{
biomes
}
\description{
A simplified and packed SpatVector of the polygons of the biomes present in
Brazilian territory. The spatial data was originally obtained from
\code{geobr::read_biomes}. Borders have been simplified by removing vertices
of borders using \code{terra::simplifyGeom}. It's necessary unpack the
Spatvectos using \code{terra::unwrap}

@usage data(biomes)
biomes <- terra::unwrap(biomes)
}
\keyword{datasets}
