% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occurrences}
\alias{occurrences}
\title{Records of plant species}
\format{
A \code{data.frame} with 1521 rows and 3 variables:
\describe{
\item{species}{Species names (Araucaria angustifolia, Abatia americana, Passiflora edmundoi, Myrcia hatschbachii, Serjania pernambucensis, Inga virescens,  and Solanum restingae)}
\item{x}{Longitude}
\item{y}{Latitude}
}
}
\usage{
data(occurrences)
}
\description{
A dataset containing records of 7 plant species downloaded from GBIF. The records were obtained with \code{plantR::rgbif2}
}
\keyword{datasets}
