% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{SpatVector of the federal states of Brazil}
\format{
A \code{SpatVector} with 27 geometries and 3 attributes:
\describe{
\item{abbrev_state}{State acronym}
\item{name_state}{State's full name}
\item{name_region}{The region to which the state belongs}
}
}
\usage{
states
}
\description{
A simplified and packed SpatVector of the polygons of the federal states of
Brazil. The spatial data was originally obtained from
\code{geobr::read_state}. Borders have been simplified by removing vertices
of borders using \code{terra::simplifyGeom}. It's necessary unpack the
Spatvectos using \code{terra::unwrap}

@usage data(states)
states <- terra::unwrap(states)
}
\keyword{datasets}
