% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bf_data}
\alias{bf_data}
\title{Flora e Funga do Brasil database - Version 393.401}
\format{
A \code{data.frame} with 50010 rows and 23 variables:
\describe{
\item{species}{Species names}
\item{scientificName}{Complete scientific name of the species}
\item{acceptedName}{Accepted name of the species (NA when the name in
species is already an accepted name)}
\item{kingdom}{Kingdom to which species belongs (Plantae or Fungi)}
\item{group}{Major group to which species belongs (Angiosperms,
Gymnosperms, Ferns and Lycophytes, Bryophytes, and Algae)}
\item{subgroup}{Subgroup to which species belongs. Only available for
Bryophytes (Mosses, Hornworts, and Liverworts)}
\item{phylum}{Phylum to which species belongs}
\item{class}{Class to which species belongs}
\item{order}{Order to which species belongs}
\item{family}{Family to which species belongs}
\item{genus}{Genus to which species belongs}
\item{lifeForm}{Life form of the species (e.g: Tree, Herb, Shrub, etc.)}
\item{habitat}{Habitat type of the species (e.g., Terrestrial, Rupicolous,
Epiphytic, etc.)}
\item{biome}{Biomes with confirmed occurrences of the species}
\item{states}{Federal states with confirmed occurrences of the species}
\item{vegetation}{Vegetation types with confirmed occurrences of the
species}
\item{origin}{Indicates whether the species is Native, Naturalized, or
Cultivated in Brazil}
\item{endemism}{Indicates whether the species is Endemic or Non-endemic to
Brazil}
\item{taxonomicStatus}{Indicates the level of recognition and acceptance
of the species (Accepted or Synonym)}
\item{nomenclaturalStatus}{Indicates the legitimacy and validity of the
species name (Correct, Illegitimate, Uncertain_Application, etc.)}
\item{vernacularName}{Locally or culturally used name for the species}
\item{taxonRank}{Taxonomic rank (Species, Genus, Family, Order, etc). This
data contains only Species}
\item{id}{Unique id for species}
}
}
\usage{
data(bf_data)
}
\description{
A dataset containing a subset of the Flora e Funga do Brasil database
(version 393.401)
}
\references{
Flora e Funga do Brasil. Jardim Botânico do Rio de Janeiro. Available at:
http://floradobrasil.jbrj.gov.br/
}
\keyword{datasets}
